% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Dsp-Package.R
\docType{package}
\name{dspr}
\alias{dspr}
\alias{dspr-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

