% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Format-GetSamplerObj.R
\name{getSamplerObj}
\alias{getSamplerObj}
\title{DSP input data structure creation}
\usage{
getSamplerObj(modelObj, cycVec, fwLen, useNA)
}
\arguments{
\item{modelObj}{stuff}

\item{cycVec}{stuff}

\item{fwLen}{stuff}

\item{useNA}{stuff}
}
\value{
\code{getSamplerObj} returns a \code{list} containing the following
  components:

  \describe{

  \item{\code{U}}{Design matrix of model covariates.  If any baseline- and /
  or cycle-specific covariates are included in the model, then they are
  expanded to a daily format.  Categorical variables, if any, are encoded in
  dummy-format coding.

  The rows of U are sorted first by subject id and then by cycle. Observation
  days in which intercourse did not occur are removed.  **** More to come..
  }

  \item{\code{idDayExpan}}{}

  \item{\code{gamBinBool}}{}

  \item{\code{varNames}}{}

  \item{\code{useNaSexBool}}{}

  \item{\code{subjId}}{}

  \item{\code{n}}{}

  \item{\code{nObs}}{}

  \item{\code{q}}{}

  \strong{idx}: if \code{useNA} is specified as \code{"none"}, then a
  \code{list} with name \code{idx} is included as an object

  \describe{

  \item{\code{preg}}{****}

  \item{\code{U}}{****}

  \item{\code{pregCyc}}{****}

  \item{\code{subj}}{****}

  }

  \strong{naSexDat}: if \code{useNA} is specified as \code{"none"},
  then the following object is included as in element in the return
  \code{list} object:

  }
}
\description{
\code{getSamplerObj} creates..
}
\details{
Add details about the software
}

